separated results into blueprint as well as session check is_logged into separate module. made edits to some html files to app routing
