+++ application/templates/myTests.html
@@ -37,7 +37,7 @@
 
             <td>10</td>
 
-            <td><a href="quiz-2.html">Start Quiz</a></td>
+            <td><a href="quiz-2">Start Quiz</a></td>
         </tr>
 
         <tr>
@@ -50,7 +50,7 @@
 
             <td>10</td>
 
-            <td><a href="quiz-2.html">Start Quiz</a></td>
+            <td><a href="quiz-2">Start Quiz</a></td>
 
         </tr>
 
@@ -64,7 +64,7 @@
 
             <td>10</td>
 
-            <td><a href="quiz-2.html">Start Quiz</a></td>
+            <td><a href="quiz-2">Start Quiz</a></td>
 
         </tr>
 
@@ -78,7 +78,7 @@
 
             <td>10</td>
 
-            <td><a href="quiz-2.html">Start Quiz</a></td>
+            <td><a href="quiz-2">Start Quiz</a></td>
 
         </tr>
 
@@ -92,7 +92,7 @@
 
             <td>10</td>
 
-            <td><a href="quiz-2.html">Start Quiz</a></td>
+            <td><a href="quiz-2">Start Quiz</a></td>
 
         </tr>
 
@@ -106,7 +106,7 @@
 
             <td>10</td>
 
-            <td><a href="quiz-2.html">Start Quiz</a></td>
+            <td><a href="quiz-2">Start Quiz</a></td>
 
         </tr>
