+++ application/templates/studDash.html
@@ -30,7 +30,7 @@
     <div class="superBox" style="display: flex;">
         <div class="side-menu">
             <div class="brand-name">
-                <h1><a href="index.html">QuizNose</a></h1>
+                <h1><a href="index">QuizNose</a></h1>
                 <img src="../static/avatars/3.png" alt="" class="user-photo">
             </div>
             <ul>
@@ -84,7 +84,7 @@
                 <div class="card" onclick="window.location.href = 'quiz';">
                     <div class="box">
                         <h1>10</h1>
-                        <h3><a href="myTests.html">My Tests</a></h3>
+                        <h3><a href="myTests">My Tests</a></h3>
                     </div>
                     <div class="icon-case">
                         <img src="../static/avatars/lib.jpg" alt="">
